+++ changelog.md
@@ -13,6 +13,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Add sided subscripts for [`negate ¯`](https://uiua.org/docs/negate), which get the conjugate or negative conjugate
 - Add numeric subscripts for [`rise ⍏`](https://uiua.org/docs/rise) and [`fall ⍖`](https://uiua.org/docs/fall), which give [`pick ⊡`](https://uiua.org/docs/pick)-compatible indices
 - Lone numeric subscripts now format to number literals, making it easier to split them off
+- Add `NoInit` constructor to data definitions construct data without running initializers (but still running validators)
 - Allow for function arguments to macros to be indicated in a signature comment
   - Put signature comments for each function between `()`s at the beginning of the macro's signature comment
 - Add experimental `ᵢ` and `ᵣ` subscripts, which format from `,i` and `,r` respectively